verifyai/requirements.txt
transformers
torch
numpy
pandas
opencv-python-headless
Pillow
pdfplumber
docx2txt
requests
tqdm
protobuf
PyYAML
typing_extensions
